+++ advault/main.py
@@ -4,6 +4,7 @@ from fastapi import FastAPI
 from db.database import *
 from db.crud import *
 from db.models import *
+from db.schema import *
 
 app = FastAPI()
 
@@ -17,18 +18,7 @@ async def create():
 @app.post("/upload")
 async def upload(data: Aadhaar):
     ins_data_hub(data)
-
     return {"message": "Data uploaded"}
-
-
-@app.get("/download/{refrence_id}")
-async def download(refrence_id: str):
-    pass
-
-
-@app.get("/download/{refrence_id}/{file_name}")
-async def download(refrence_id: str, file_name: str):
-    pass
 
 
 @app.get("/download_all/")
